+++ refactor-me.Tests/refactor-me.Tests.csproj
@@ -63,8 +63,9 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
+    <Compile Include="Integration\TestHelpers.cs" />
+    <Compile Include="Integration\ProductOptionsControllerTests.cs" />
     <Compile Include="Integration\ProductsControllerTests.cs" />
-    <Compile Include="UnitTest1.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
